pyproject.toml
[tool.poetry]
name = "TradingBotCommons"
version = "0.0.0"
description = "TradingBotCommons is a package that contains Trading Bot common code."
authors = ["Federico Pereyra López"]
license = "MIT"
readme = "README.rst"
homepage = "https://github.com/Luxoimum/TradingBotCommons"
repository = "https://github.com/Luxoimum/TradingBotCommons"
packages = [
  { include = "src/**/*" },
  { include = "src" },
  { include = "test/**/*" },
  { include = "test" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
toml = "^0.10.2"
boto3 = "^1.26.60"
moto = "^4.1.1"
python-binance = "^1.0.17"
numpy = "^1.24.2"
pandas = "^2.1.1"
pandas-ta = "^0.3.14b0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pylint = "^2.4.0"
pytest = "^7.1.1"
pytest-cov = "^4.0.0"
pytest-runner = "^6.0.0"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry>=1.2"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
norecursedirs = [
    "node_modules"
]
